feat: promote analytics v2 route
